configs/kueue/kueuer/pyproject.toml
[project]
name = "kueuer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Shiny Brar", email = "shiny.brar@nrc-cnrc.gc.ca" }
]
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "httpx>=0.28.1",
    "kubernetes>=32.0.1",
    "kubernetes-typed[client]>=18.20.2",
    "logfire[httpx]>=3.7.1",
    "nest-asyncio>=1.6.0",
    "pandas>=2.2.3",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.2",
    "seaborn>=0.13.2",
    "typer>=0.15.2",
]
scripts = { kr = "kueuer.cli:app" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
select = ["E", "F", "W", "C", "N", "B", "I"]
ignore = ["B904", "B028"]
extend-ignore = ["D"]

[tool.ruff.lint.isort]
combine-as-imports = true

[dependency-groups]
dev = [
    "ipython>=9.0.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
